+++ loggo/loggo.py
@@ -10,7 +10,7 @@ import uuid
 from contextlib import contextmanager
 from datetime import datetime
 from functools import wraps
-from typing import Optional, Set, Dict, Union, Callable, Generator, Any, cast, Mapping, Tuple
+from typing import Optional, Set, Dict, Union, Callable, Generator, Any, cast, Mapping
 
 # you don't need graylog installed
 try:
@@ -41,17 +41,17 @@ class Loggo:
                  returned: Optional[str] = FORMS['returned'],
                  returned_none: Optional[str] = FORMS['returned_none'],
                  errored: Optional[str] = FORMS['errored'],
-                 error_level: int = 50,
+                 error_level: int = logging.INFO,
                  facility: str = 'loggo',
                  ip: Optional[str] = None,
-                 port: Optional[str] = None,
+                 port: Optional[int] = None,
                  do_print: bool = False,
                  do_write: bool = False,
                  truncation: int = 7500,
                  raise_logging_errors: bool = False,
-                 logfile: Optional[str] = './logs/logs.txt',
+                 logfile: str = './logs/logs.txt',
                  obscured: Optional[str] = '********',
-                 private_data: Set[str] = set(),
+                 private_data: Optional[Set[str]] = None,
                  max_dict_depth: int = 5,
                  log_if_graylog_disabled: bool = True) -> None:
         """
@@ -88,21 +88,21 @@ class Loggo:
         self.raise_logging_errors = raise_logging_errors
         self.logfile = logfile
         self.obscured = obscured
-        self.private_data = private_data
+        self.private_data = private_data or set()
         self.max_dict_depth = max_dict_depth
         self.log_if_graylog_disabled = log_if_graylog_disabled
         self.logger = logging.getLogger(self.facility)  # pylint: disable=no-member
         self.logger.setLevel(Loggo.log_threshold)
         self._add_graylog_handler()
 
-    def _best_returned_none(self, returned, returned_none):
+    def _best_returned_none(self, returned: Optional[str], returned_none: Optional[str]) -> Optional[str]:
         """
         If the user has their own msg format for 'returned' logs, but not one
         for 'returned_none', we should use theirs over loggo's default
         """
         # if the user explicitly doesn't want logs for returns, set to none
         if not returned_none or not returned:
-            return
+            return None
         # if they provided their own, use that
         if returned_none != FORMS['returned_none']:
             return returned_none
@@ -278,22 +278,21 @@ class Loggo:
 
             # 'called' log tells you what was called and with what arguments
             if not just_errors:
-                msg = self.instantiated if is_init else self.called
-                self._generate_log(msg, None, formatters, param_strings)
+                self._generate_log(where, None, formatters, param_strings)
 
             try:
                 # where the original function is actually run
                 response = function(*args, **kwargs)
-                msg = self.returned if response is not None else self.returned_none
+                where = 'returned_none' if response is None else 'returned'
                 # the successful return log
                 if not just_errors and not is_init:
-                    self._generate_log(msg, response, formatters, param_strings)
+                    self._generate_log(where, response, formatters, param_strings)
                 # return whatever the original callable did
                 return response
             # handle any possible error
             except Exception as error:
                 formatters['traceback'] = traceback.format_exc()
-                self._generate_log(self.errored, error, formatters, param_strings)
+                self._generate_log('errored', error, formatters, param_strings)
                 raise
         return full_decoration
 
@@ -310,7 +309,7 @@ class Loggo:
         return params
 
     @staticmethod
-    def _make_call_signature(function: Callable, param_strings: Dict[str, str]) -> Tuple[str, Dict]:
+    def _make_call_signature(function: Callable, param_strings: Dict[str, str]) -> Dict:
         """
         Represent the call as a string mimicking how it is written in Python.
 
@@ -386,7 +385,7 @@ class Loggo:
 
         return '({})'.format(self._force_string_and_truncate(response, truncate, use_repr=True))
 
-    def _generate_log(self, msg: str, returned: Any, formatters: Dict, safe_log_data: Dict[str, str]) -> None:
+    def _generate_log(self, where: str, returned: Any, formatters: Dict, safe_log_data: Dict[str, str]) -> None:
         """
         generate message, level and log data for automated logs
 
@@ -395,12 +394,17 @@ class Loggo:
         formatters (dict): dict containing format strings needed for message
         safe_log_data (dict): dict of stringified, truncated, censored parameters
         """
+        # if the user turned off logs of this type, do nothing immediately
+        msg = getattr(self, where)
+        if not msg:
+            return
+
         # if errors not to be shown and this is an error, quit
-        if not self.allow_errors and msg == self.errored:
+        if not self.allow_errors and where == 'errored':
             return
 
         # if state is stopped and not an error, quit
-        if self.stopped and msg != self.errored:
+        if self.stopped and where != 'errored':
             return
 
         # do not log loggo, because why would you ever want that?
@@ -408,18 +412,18 @@ class Loggo:
             return
 
         # return value for log message
-        if msg == self.returned:
+        if 'returned' in where:
             ret_str = self._represent_return_value(returned, truncate=None)
             formatters['return_value'] = ret_str
             formatters['return_type'] = type(returned).__name__
 
         # if what is 'returned' is an exception, get the error formatters
-        if msg == self.errored:
+        if where == 'errored':
             formatters['exception_type'] = type(returned).__name__
             formatters['exception_msg'] = str(returned)
             formatters['level'] = self.error_level
         else:
-            formatters['level'] = 20
+            formatters['level'] = logging.INFO
 
         # format the string template
         msg = msg.format(**formatters).replace('  ', ' ')
@@ -469,8 +473,8 @@ class Loggo:
         """
         try:
             obj = str(obj) if not use_repr else repr(obj)
-        except Exception as error:
-            self.warning('Object could not be cast to string', extra=dict(exception_type=type(error), error=error))
+        except Exception as exc:
+            self.warning('Object could not be cast to string', extra=dict(exception_type=type(exc), exception=exc))
             return '<<Unstringable input>>'
         if truncate is None:
             return obj
@@ -528,16 +532,16 @@ class Loggo:
             extra = self.sanitise(extra, use_repr=False)
             msg = self.sanitise_msg(msg)
 
-        extra.update(dict(level=str(level), loggo=True))
+        extra.update(dict(level=str(level), loggo=str(True)))
 
-        # print or write log lines
-        trace = extra.get('traceback')
+        # format logs for printing/writing to file
         if self.do_write or self.do_print:
-            now = datetime.now().strftime('%d.%m %Y %H:%M:%S')
-            timestamp = extra.get('timestamp', now)
-            line = f'{now}\t{msg}\t{level}'
-        if trace:
-            line = f'{msg} -- see below: \n{trace}\n' if trace else msg
+            ts = extra.get('timestamp', datetime.now().strftime('%d.%m %Y %H:%M:%S'))
+            line = f'{ts}\t{msg}\t{level}'
+            trace = extra.get('traceback')
+            if trace:
+                line = f'{line} -- see below: \n{trace}\n'
+        # do printing and writing to file
         if self.do_print:
             print(line)
         if self.do_write: